Please note that some changes were necessary to the GC as well, and I am trusting that Dave Friedman got things right.  Again, most of these were changes from uint to size_t, but things here seem a bit less clear-cut than most of the changes to the GDC runtime itself.

Finally, it is worth being aware that of the GC code, only gcx.d is really pretty much the same between Tango and Phobos/GPhobos.  The remaining files are split among the GC directory (gc/basic), and the files lifetime.d and memory.d in the compiler runtime code.  The Tango design is far cleaner, but the structural differences complicate merging changes in this area.  I've triple-checked all changes for this merge, but if there are any functional problems, I expect them to concern these GC-related files.

+++ lib/compiler/dmd/arraycat.d
@@ -37,7 +37,7 @@ private
 
 extern (C):
 
-byte[] _d_arraycopy(uint size, byte[] from, byte[] to)
+byte[] _d_arraycopy(size_t size, byte[] from, byte[] to)
 {
     debug printf("f = %p,%d, t = %p,%d, size = %d\n",
                  cast(void*)from, from.length, cast(void*)to, to.length, size);
